remove USE_LDCONFIG and add BUNDLE_LIBS
Shared libraries installed along with vscode are for internal use and
not intended for public consumption.

Remove USE_LDCONFIG so as not to add the shared library directory to
search paths. Add BUNDLE_LIBS so that pkg(8) does not pick the
libraries up for packaging.

Related GitHub Issue: https://github.com/freebsd/pkg/issues/2368