Update to 1.102.0
While here, remove nearly useless checksum verification.

The checksum test is for checking integrity of a downloaded zip
file. In our case, the zip file is generated on-the-fly from an
electron installation instead of donwloading. So making checksum from
the generated zip file and checking it with the same zip file does not
make much sense.

Changelog: https://code.visualstudio.com/updates/v1_102